+++ pjmedia/src/pjmedia/sdp.c
@@ -640,7 +640,7 @@ static int print_media_desc( pjmedia_sdp_media *m, char *buf, int len)
 
     /* print connection info, if present. */
     if (m->conn) {
-	printed = print_connection_info(m->conn, p, end-p);
+	printed = print_connection_info(m->conn, p, (int)(end-p));
 	if (printed < 0) {
 	    return -1;
 	}
@@ -649,7 +649,7 @@ static int print_media_desc( pjmedia_sdp_media *m, char *buf, int len)
     
     /* print optional bandwidth info. */
     for (i=0; i<m->bandw_count; ++i) {
-	printed = print_bandw(m->bandw[i], p, end-p);
+	printed = (int)print_bandw(m->bandw[i], p, end-p);
 	if (printed < 0) {
 	    return -1;
 	}
@@ -658,14 +658,14 @@ static int print_media_desc( pjmedia_sdp_media *m, char *buf, int len)
 
     /* print attributes. */
     for (i=0; i<m->attr_count; ++i) {
-	printed = print_attr(m->attr[i], p, end-p);
+	printed = (int)print_attr(m->attr[i], p, end-p);
 	if (printed < 0) {
 	    return -1;
 	}
 	p += printed;
     }
 
-    return p-buf;
+    return (int)(p-buf);
 }
 
 PJ_DEF(pjmedia_sdp_media*) pjmedia_sdp_media_clone(
@@ -805,7 +805,7 @@ static int print_session(const pjmedia_sdp_session *ses,
 
     /* Connection line (c=) if exist. */
     if (ses->conn) {
-	printed = print_connection_info(ses->conn, p, end-p);
+	printed = print_connection_info(ses->conn, p, (int)(end-p));
 	if (printed < 1) {
 	    return -1;
 	}
@@ -814,7 +814,7 @@ static int print_session(const pjmedia_sdp_session *ses,
 
     /* print optional bandwidth info. */
     for (i=0; i<ses->bandw_count; ++i) {
-	printed = print_bandw(ses->bandw[i], p, end-p);
+	printed = (int)print_bandw(ses->bandw[i], p, end-p);
 	if (printed < 1) {
 	    return -1;
 	}
@@ -837,7 +837,7 @@ static int print_session(const pjmedia_sdp_session *ses,
 
     /* Print all attribute (a=) lines. */
     for (i=0; i<ses->attr_count; ++i) {
-	printed = print_attr(ses->attr[i], p, end-p);
+	printed = (int)print_attr(ses->attr[i], p, end-p);
 	if (printed < 0) {
 	    return -1;
 	}
@@ -846,14 +846,14 @@ static int print_session(const pjmedia_sdp_session *ses,
 
     /* Print media (m=) lines. */
     for (i=0; i<ses->media_count; ++i) {
-	printed = print_media_desc(ses->media[i], p, end-p);
+	printed = print_media_desc(ses->media[i], p, (int)(end-p));
 	if (printed < 0) {
 	    return -1;
 	}
 	p += printed;
     }
 
-    return p-buf;
+    return (int)(p-buf);
 }
 
 /******************************************************************************
