+++ pjnath/src/pjnath-test/concur_test.c
@@ -72,7 +72,7 @@ static int server_thread_proc(void *p)
 	/* Serve client */
 	PJ_FD_ZERO(&rdset);
 	PJ_FD_SET(test_sess->server_sock, &rdset);
-	n = pj_sock_select(test_sess->server_sock+1, &rdset,
+	n = pj_sock_select((int)test_sess->server_sock+1, &rdset,
 	                   NULL, NULL, &timeout);
 	if (n==1 && PJ_FD_ISSET(test_sess->server_sock, &rdset)) {
 	    pj_uint8_t pkt[512];
