+++ pjnath/src/pjnath/ice_strans.c
@@ -438,7 +438,7 @@ static pj_status_t create_comp(pj_ice_strans *ice_st, unsigned comp_id)
 				   cand->type, &cand->base_addr);
 
 	    /* Set default candidate to srflx */
-	    comp->default_cand = cand - comp->cand_list;
+	    comp->default_cand = (unsigned)(cand - comp->cand_list);
 
 	    pj_log_pop_indent();
 	}
@@ -1197,14 +1197,17 @@ PJ_DEF(pj_status_t) pj_ice_strans_sendto( pj_ice_strans *ice_st,
 		comp->turn_log_off = PJ_TRUE;
 	    }
 
-	    status = pj_turn_sock_sendto(comp->turn_sock, (const pj_uint8_t*)data, data_len,
+	    status = pj_turn_sock_sendto(comp->turn_sock, 
+					 (const pj_uint8_t*)data, 
+					 (unsigned)data_len,
 					 dst_addr, dst_addr_len);
 	    return (status==PJ_SUCCESS||status==PJ_EPENDING) ? 
 		    PJ_SUCCESS : status;
 	} else {
 	    pkt_size = data_len;
 	    status = pj_stun_sock_sendto(comp->stun_sock, NULL, data, 
-					 data_len, 0, dst_addr, dst_addr_len);
+					 (unsigned)data_len, 0, dst_addr, 
+					 dst_addr_len);
 	    return (status==PJ_SUCCESS||status==PJ_EPENDING) ? 
 		    PJ_SUCCESS : status;
 	}
@@ -1339,14 +1342,15 @@ static pj_status_t ice_tx_pkt(pj_ice_sess *ice,
     if (transport_id == TP_TURN) {
 	if (comp->turn_sock) {
 	    status = pj_turn_sock_sendto(comp->turn_sock, 
-					 (const pj_uint8_t*)pkt, size,
+					 (const pj_uint8_t*)pkt, 
+					 (unsigned)size,
 					 dst_addr, dst_addr_len);
 	} else {
 	    status = PJ_EINVALIDOP;
 	}
     } else if (transport_id == TP_STUN) {
 	status = pj_stun_sock_sendto(comp->stun_sock, NULL, 
-				     pkt, size, 0,
+				     pkt, (unsigned)size, 0,
 				     dst_addr, dst_addr_len);
     } else {
 	pj_assert(!"Invalid transport ID");
@@ -1520,7 +1524,7 @@ static pj_bool_t stun_on_status(pj_stun_sock *stun_sock,
 
 		if (dup) {
 		    /* Duplicate found, remove the srflx candidate */
-		    unsigned idx = cand - comp->cand_list;
+		    unsigned idx = (unsigned)(cand - comp->cand_list);
 
 		    /* Update default candidate index */
 		    if (comp->default_cand > idx) {
@@ -1563,7 +1567,7 @@ static pj_bool_t stun_on_status(pj_stun_sock *stun_sock,
 			  comp->comp_id));
 
 		if (cand) {
-		    unsigned idx = cand - comp->cand_list;
+		    unsigned idx = (unsigned)(cand - comp->cand_list);
 
 		    /* Update default candidate index */
 		    if (comp->default_cand == idx) {
@@ -1694,7 +1698,7 @@ static void turn_on_state(pj_turn_sock *turn_sock, pj_turn_state_t old_state,
 	cand->status = PJ_SUCCESS;
 
 	/* Set default candidate to relay */
-	comp->default_cand = cand - comp->cand_list;
+	comp->default_cand = (unsigned)(cand - comp->cand_list);
 
 	PJ_LOG(4,(comp->ice_st->obj_name, 
 		  "Comp %d: TURN allocation complete, relay address is %s",
